+++ inc/social/rank-system/chill-forums-rank.php
@@ -27,7 +27,7 @@ function extrachill_add_rank_and_points_to_reply() {
 	$local_scene = extrachill_community_get_public_local_scene( $reply_author_id );
 	if ( ! empty( $local_scene['name'] ) ) {
 		echo '<div class="reply-author-local-scene">';
-		echo '<span>Local Scene:</span> ' . esc_html( $local_scene['name'] );
+		extrachill_community_render_local_scene_badge( $local_scene, true );
 		echo '</div>';
 	}